Fixed project structure. Renamed fxl to proper format.
