+++ src/main/java/Controller/SimulationController.java
@@ -1,4 +1,4 @@
-package View;
+package Controller;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
